==> Steps_taken/2-Fill in Missing Values/1 - Fill ZIP.ipynb <==
# Project 1

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy import stats

data = pd.read_csv('NY property data.csv')
data.head()

data.info()

## Fill in Field ZIP

data[data['ZIP'].isnull()].head()

### Fill in missing values by B and BLOCK 

# select record which has null in ZIP
zip = data[data['ZIP'].isnull()][['B','BLOCK','STADDR','ZIP']]
zip['RECORD'] = zip.index + 1 
zip.head()

# select record with value in ZIP
data_withzip = data[data['ZIP'].notnull()]
data_withzip.head()

# group record with value in ZIP by B and BLOCK
zipfill = data_withzip.groupby(['B','BLOCK'])[['ZIP']].agg(lambda x: stats.mode(x)[0][0]).reset_index()
zipfill.head()

# Fill in the missing value using the most common ZIP value of records with same B and BLOCK
zip1 = zip.merge(zipfill, on = ['B','BLOCK'], how = 'left')
print(sum(zip1['ZIP_y'].isnull()))
print(sum(zip1['ZIP_y'].notnull()))

zip1.head()

zip1 = zip1[['RECORD','ZIP_y']]
zip1.head()

## left join the original table and table with filled value by RECORD
data = data.merge(zip1, on = ['RECORD'], how = 'left')
data.head()

data.ZIP = data.ZIP.fillna(data['ZIP_y'])
del data['ZIP_y']
data.head()

data[['ZIP']].info()

### Filling in missing values by B and STADDR group 

# select record which has null in ZIP
zip = data[data['ZIP'].isnull()][['B','BLOCK','STADDR','ZIP']]
zip['RECORD'] = zip.index + 1 
zip.head()

zipfill2 = data_withzip.groupby(['B','STADDR'])[['ZIP']].agg(lambda x: stats.mode(x)[0][0]).reset_index()
zipfill2.head()

zip2 = zip.merge(zipfill2, on = ['B','STADDR'], how = 'left')
del zip2['ZIP_x']
print(sum(zip2['ZIP_y'].isnull()))
print(sum(zip2['ZIP_y'].notnull()))
zip2.head()

zip2 = zip2[['RECORD','ZIP_y']]
zip2.head()

data = data.merge(zip2, on = ['RECORD'], how = 'left')
data.head()

data.ZIP = data.ZIP.fillna(data['ZIP_y'])
data.head()

del data['ZIP_y']
data.head()

data[['ZIP']].info()

### Fill in Missing Values by B 

# select record which has null in ZIP
zip = data[data['ZIP'].isnull()][['B','BLOCK','STADDR','ZIP']]
zip['RECORD'] = zip.index + 1 
zip.head()

zipfill3 = data_withzip.groupby(['B'])[['ZIP']].agg(lambda x: stats.mode(x)[0][0]).reset_index()
zipfill3.head()

zip3 = zip.merge(zipfill3, on = ['B'], how = 'left')
zip3 = zip3[['RECORD','ZIP_y']]
zip3.head()

data = data.merge(zip3, on = ['RECORD'], how = 'left')
data.ZIP = data.ZIP.fillna(data['ZIP_y'])
data.head()

del data['ZIP_y']
data[['ZIP']].info()

data.to_csv('NY property data_zip.csv')

mydata = pd.read_csv('NY property data_zip.csv')
mydata.info()
