==> Steps_taken/2-Fill in Missing Values/2 - Fill Stories.ipynb <==
Project 1

import pandas as pd
import numpy as np
import scipy.stats as sps
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn as skl
%matplotlib inline

%%time
mydata=pd.read_csv('NY property data_zip.csv')

mydata.head()

df = pd.DataFrame(mydata)

df.STORIES.isna().sum()

Fill in missing values by TAXCLASS and ZIP

stories = mydata[mydata['STORIES'].isnull()][['RECORD','TAXCLASS','ZIP','BLOCK','STORIES']]

stories

data_withstories = mydata[mydata['STORIES'].notnull()]
data_withstories.head()

group1=data_withstories.groupby(['TAXCLASS', 'ZIP']).agg({'STORIES':'mean','BBLE':'count'})
group1=group1.reset_index(['TAXCLASS','ZIP'])
group1.head()

groupfil1=group1.query('BBLE>=10')
groupfil1.head()

stories=stories.merge(groupfil1, on =['TAXCLASS','ZIP'], how = 'left')
del stories['STORIES_x']
stories

stories1 = stories[stories['STORIES_y'].isnull()][['RECORD','TAXCLASS','ZIP','BLOCK','STORIES_y']]

stories1

group2=data_withstories.groupby(['BLOCK']).agg({'STORIES':'mean','BBLE':'count'})

group2.head()

groupfil2=group2.query('BBLE>=5')
groupfil2.head()

stories2=stories1.merge(groupfil2, on =['BLOCK'], how = 'left')
del stories2['STORIES_y']

stories2.head()

stories3=stories.merge(stories2, on =['RECORD'], how = 'left')

stories3.STORIES_y = stories3.STORIES_y.fillna(stories3['STORIES'])

del stories3['BBLE_x']
del stories3['TAXCLASS_y']
del stories3['ZIP_y']
del stories3['BLOCK_y']
del stories3['STORIES']
del stories3['BBLE_y']

stories3.head()

stories4 = stories3[stories3['STORIES_y'].isnull()][['RECORD','TAXCLASS_x','ZIP_x','BLOCK_x','STORIES_y']]

stories4

group3=data_withstories.groupby(['ZIP']).agg({'STORIES':'mean'})

group3.head()

stories5=stories4.merge(group3, left_on =['ZIP_x'], right_on=['ZIP'],how = 'left')
del stories5['STORIES_y']
stories5


storiesfinal=stories3.merge(stories5, on =['RECORD'], how = 'left')

storiesfinal.STORIES_y = storiesfinal.STORIES_y.fillna(storiesfinal['STORIES'])



del storiesfinal['TAXCLASS_x_y']
del storiesfinal['ZIP_x_y']
del storiesfinal['BLOCK_x_y']
del storiesfinal['STORIES']

storiesfinal.head()

storiesdata=mydata.merge(storiesfinal, on =['RECORD'], how = 'left')

storiesdata.STORIES = storiesdata.STORIES.fillna(storiesdata['STORIES_y'])


del storiesdata['Unnamed: 0']
del storiesdata['TAXCLASS_x_x']
del storiesdata['ZIP_x_x']
del storiesdata['BLOCK_x_x']
del storiesdata['STORIES_y']

storiesdata.head()

df = pd.DataFrame(storiesdata)

df.STORIES.isna().sum()

storiesdata.to_csv('NY property data_zip_stories.csv')

data=pd.read_csv('NY property data_zip_stories.csv')

data['STORIES'].describe()
