==> Steps_taken/2-Fill in Missing Values/3 - Fill 7 Fields.ipynb <==
import pandas as pd
import numpy as np
import scipy.stats as sps
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn as skl
from sklearn import preprocessing
import math
%matplotlib inline

%%time
mydata = pd.read_csv('NY property data_zip_stories.csv')

1. Refill the FULLVAL, AVLAND, AVTOT, add new classifier: the maximum of (LTFRONT, LTDEPTH, BLDFRONT and BLDDEPTH) 

2. Refill the LTFRONT, LTDEPTH, BLDFRONT, BLDDEPTH, add new classifier: the bin_value of the mean of (FULLVAL, AVLAND, AVTOT)


一、FULLVAL

Turn 0s into NAN

mydata['FULLVAL']=mydata['FULLVAL'].replace(0, np.nan)
mydata['FULLVAL'].isna().value_counts()

Turn STORIES to INT

data = mydata
data['TAXCLASS']=mydata['TAXCLASS'].astype('str')
data['ZIP']=mydata['ZIP'].astype('str')
data['STORIES']=mydata['STORIES'].astype(np.float64)

data['STORIES']=round(data['STORIES'])
data['STORIES']=mydata['STORIES'].astype('str')

Consider the maximum of (LTFRONT, LTDEPTH, BLDFRONT and BLDDEPTH) as classifier

data['new'] = data[['LTFRONT','LTDEPTH','BLDFRONT','BLDDEPTH']].max(axis=1)
bins = [i*50 for i in range(201)]
groups = [i for i in range(200)]
data['area_level'] = pd.cut(data['new'], bins, labels=groups)
data['area_level']=data['area_level'].replace(np.nan,0)
data['area_level']=mydata['area_level'].astype('str')

use TAXCLASS STORIES area_level ZIP


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['FULLVAL'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanFULLVAL'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanFULLVAL':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanFULLVAL_y']].groupby('group')['meanFULLVAL_y'].mean()
del data10['meanFULLVAL_x']

mydata['FULLVAL'].fillna(data10['meanFULLVAL_y'],inplace=True ) 

mydata['FULLVAL'].isna().value_counts()


Under 10 use TAXCLASS STORIES area_level


data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['FULLVAL'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanFULLVAL'] = data['group'].apply(get_matched_value)

dt1= data.groupby(['group']).agg({'meanFULLVAL':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanFULLVAL_y']].groupby('group')['meanFULLVAL_y'].mean()
del data10['meanFULLVAL_x']

mydata['FULLVAL'].fillna(data10['meanFULLVAL_y'],inplace=True ) 

mydata['FULLVAL'].isna().value_counts()


Under 10 use STORIES,area_level

data['group'] = data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['FULLVAL'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanFULLVAL'] = data['group'].apply(get_matched_value)

dt1= data.groupby(['group']).agg({'meanFULLVAL':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanFULLVAL_y']].groupby('group')['meanFULLVAL_y'].mean()
del data10['meanFULLVAL_x']

mydata['FULLVAL'].fillna(data10['meanFULLVAL_y'],inplace=True ) 

mydata['FULLVAL'].isna().value_counts()

STILL HAVE NULL —— use stories

data['group'] = data['STORIES']
table = data.groupby(['group'])['FULLVAL'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanFULLVAL'] = data['group'].apply(get_matched_value)

dt1= data.groupby(['group']).agg({'meanFULLVAL':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanFULLVAL_y']].groupby('group')['meanFULLVAL_y'].mean()
del data10['meanFULLVAL_x']

mydata['FULLVAL'].fillna(data10['meanFULLVAL_y'],inplace=True ) 

mydata['FULLVAL'].isna().value_counts()

STILL HAVE NULL —— use TAXCLASS

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['FULLVAL'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanFULLVAL'] = data['group'].apply(get_matched_value)

dt1= data.groupby(['group']).agg({'meanFULLVAL':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanFULLVAL_y']].groupby('group')['meanFULLVAL_y'].mean()
del data10['meanFULLVAL_x']

mydata['FULLVAL'].fillna(data10['meanFULLVAL_y'],inplace=True ) 

mydata['FULLVAL'].isna().value_counts()

二、AVLAND

mydata['AVLAND']=mydata['AVLAND'].replace(0, np.nan)


mydata['AVLAND'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVLAND'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVLAND'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVLAND':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVLAND_y']].groupby('group')['meanAVLAND_y'].mean()
del data10['meanAVLAND_x']

mydata['AVLAND'].fillna(data10['meanAVLAND_y'],inplace=True ) 

mydata['AVLAND'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVLAND'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVLAND'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVLAND':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVLAND_y']].groupby('group')['meanAVLAND_y'].mean()
del data10['meanAVLAND_x']

mydata['AVLAND'].fillna(data10['meanAVLAND_y'],inplace=True ) 

mydata['AVLAND'].isna().value_counts()

data['group'] = data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVLAND'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVLAND'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVLAND':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVLAND_y']].groupby('group')['meanAVLAND_y'].mean()
del data10['meanAVLAND_x']

mydata['AVLAND'].fillna(data10['meanAVLAND_y'],inplace=True ) 

mydata['AVLAND'].isna().value_counts()

data['group'] = data['STORIES']
table = data.groupby(['group'])['AVLAND'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVLAND'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVLAND':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVLAND_y']].groupby('group')['meanAVLAND_y'].mean()
del data10['meanAVLAND_x']

mydata['AVLAND'].fillna(data10['meanAVLAND_y'],inplace=True ) 

mydata['AVLAND'].isna().value_counts()

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['AVLAND'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVLAND'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVLAND':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVLAND_y']].groupby('group')['meanAVLAND_y'].mean()
del data10['meanAVLAND_x']

mydata['AVLAND'].fillna(data10['meanAVLAND_y'],inplace=True ) 

mydata['AVLAND'].isna().value_counts()

三、AVTOT

mydata['AVTOT']=mydata['AVTOT'].replace(0, np.nan)

data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVTOT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVTOT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVTOT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVTOT_y']].groupby('group')['meanAVTOT_y'].mean()
del data10['meanAVTOT_x']

mydata['AVTOT'].fillna(data10['meanAVTOT_y'],inplace=True ) 

mydata['AVTOT'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVTOT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVTOT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVTOT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVTOT_y']].groupby('group')['meanAVTOT_y'].mean()
del data10['meanAVTOT_x']

mydata['AVTOT'].fillna(data10['meanAVTOT_y'],inplace=True ) 

mydata['AVTOT'].isna().value_counts()


data['group'] = data['STORIES']+'.'+data['area_level']
table = data.groupby(['group'])['AVTOT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVTOT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVTOT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVTOT_y']].groupby('group')['meanAVTOT_y'].mean()
del data10['meanAVTOT_x']

mydata['AVTOT'].fillna(data10['meanAVTOT_y'],inplace=True ) 

mydata['AVTOT'].isna().value_counts()


data['group'] = data['STORIES']
table = data.groupby(['group'])['AVTOT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVTOT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVTOT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVTOT_y']].groupby('group')['meanAVTOT_y'].mean()
del data10['meanAVTOT_x']

mydata['AVTOT'].fillna(data10['meanAVTOT_y'],inplace=True ) 

mydata['AVTOT'].isna().value_counts()

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['AVTOT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanAVTOT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanAVTOT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanAVTOT_y']].groupby('group')['meanAVTOT_y'].mean()
del data10['meanAVTOT_x']

mydata['AVTOT'].fillna(data10['meanAVTOT_y'],inplace=True ) 

mydata['AVTOT'].isna().value_counts()

Fill LTFRONT, LTDEPTH, BLDFRONT, BLDDEPTH

LTFRONT

USE value level=mean(FULLVAL, AVLAND, AVTOT) as classifier

data['meanvalue'] = data[['FULLVAL','AVLAND','AVTOT']].mean(axis=1)


data['meanvalvalue'] = data[['FULLVAL','AVLAND','AVTOT']].mean(axis=1)
bins = [i*1000 for i in range(200)]
bins.extend([a*10000+200000 for a in range(80)])
bins.extend([i*100000+1000000 for i in range(90)])
bins.extend([i*1000000+10000000 for i in range(90)])
bins.extend([i*1000000+100000000 for i in range(90)])
bins.extend([i*10000000+1000000000 for i in range(90)])
bins.extend([i*100000000+10000000000 for i in range(90)])


groups = [i for i in range(len(bins)-1)]
data['val_level'] = pd.cut(data['meanvalvalue'], bins, labels=groups)
data['val_level']=data['val_level'].replace(np.nan,0)
data['val_level']=mydata['val_level'].astype('str')

mydata['LTFRONT']=mydata['LTFRONT'].replace(0, np.nan)
mydata['LTFRONT'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

data['group'] = data['TAXCLASS']+data['val_level']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

data['group'] = data['STORIES']+data['val_level']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

data['group'] = data['val_level']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['LTFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTFRONT_y']].groupby('group')['meanLTFRONT_y'].mean()
del data10['meanLTFRONT_x']

mydata['LTFRONT'].fillna(data10['meanLTFRONT_y'],inplace=True ) 

mydata['LTFRONT'].isna().value_counts()

LTDEPTH

mydata['LTDEPTH']=mydata['LTDEPTH'].replace(0, np.nan)
mydata['LTDEPTH'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

data['group'] = data['TAXCLASS']+data['val_level']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

data['group'] = data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

data['group'] = data['val_level']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

data['group'] = data['STORIES']
table = data.groupby(['group'])['LTDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanLTDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanLTDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanLTDEPTH_y']].groupby('group')['meanLTDEPTH_y'].mean()
del data10['meanLTDEPTH_x']

mydata['LTDEPTH'].fillna(data10['meanLTDEPTH_y'],inplace=True ) 

mydata['LTDEPTH'].isna().value_counts()

BLDFRONT

mydata['BLDFRONT']=mydata['BLDFRONT'].replace(0, np.nan)
mydata['BLDFRONT'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()


data['group'] = data['TAXCLASS']+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()

data['group'] = data['TAXCLASS']+'.'data['val_level']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()

data['group'] = data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()

data['group'] =data['val_level']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['BLDFRONT'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDFRONT'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDFRONT':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=0').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDFRONT_y']].groupby('group')['meanBLDFRONT_y'].mean()
del data10['meanBLDFRONT_x']

mydata['BLDFRONT'].fillna(data10['meanBLDFRONT_y'],inplace=True ) 

mydata['BLDFRONT'].isna().value_counts()

BLDDEPTH

mydata['BLDDEPTH']=mydata['BLDDEPTH'].replace(0, np.nan)
mydata['BLDDEPTH'].isna().value_counts()


data['group'] = data['TAXCLASS']+'.'+data['ZIP']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

data['group'] = data['TAXCLASS']+'.'+data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

data['group'] = data['TAXCLASS']+'.'+data['val_level']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

data['group'] = data['STORIES']+'.'+data['val_level']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

data['group'] = data['val_level']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

data['group'] = data['TAXCLASS']
table = data.groupby(['group'])['BLDDEPTH'].mean()

table = round(table)
def get_matched_value(string):
    val = table.loc[string]
    return val


data['meanBLDDEPTH'] = data['group'].apply(get_matched_value)


dt1= data.groupby(['group']).agg({'meanBLDDEPTH':'mean','RECORD':'count'})
dt10=dt1.query('RECORD>=10').reset_index()

data10=data.merge(dt10, on =['group'], how = 'left')
data10[['group','meanBLDDEPTH_y']].groupby('group')['meanBLDDEPTH_y'].mean()
del data10['meanBLDDEPTH_x']

mydata['BLDDEPTH'].fillna(data10['meanBLDDEPTH_y'],inplace=True ) 

mydata['BLDDEPTH'].isna().value_counts()

Check unusual points in our previous result

mydata[['FULLVAL','AVLAND','AVTOT','ZIP','STORIES']][mydata['RECORD']==684704]

mydata[['FULLVAL','AVLAND','AVTOT','ZIP','STORIES']][mydata['RECORD']==776306]

mydata[['FULLVAL','AVLAND','AVTOT','ZIP','STORIES']][mydata['RECORD']==821853]

mydata[['LTFRONT','LTDEPTH','BLDDEPTH','BLDFRONT','FULLVAL','AVLAND','AVTOT','ZIP','STORIES']][mydata['RECORD']==917942]

mydata.columns


del mydata['new']
del mydata['area_level']
del mydata['group']
del mydata['meanFULLVAL']
del mydata['meanAVLAND']
del mydata['meanAVTOT']
del mydata['meanvalvalue']
del mydata['val_level']
del mydata['meanLTFRONT']
del mydata['meanLTDEPTH']
del mydata['meanBLDFRONT']
del mydata['meanBLDDEPTH']


pd.DataFrame(mydata).to_csv('NY property data_filled_Revised.csv')
